refactor: split rginx-app stage 2 modules
