+++ crates/rginx-app/src/check.rs
@@ -1,0 +1,7 @@
+mod render;
+mod routes;
+mod summary;
+mod tls;
+
+pub(crate) use render::print_check_success;
+pub(crate) use summary::build_check_summary;
