+++ scripts/modularization_baseline.json
@@ -4,8 +4,6 @@
   "test_soft_limit": 400,
   "test_hard_limit": 600,
   "legacy_production_size_ceilings": {
-    "crates/rginx-app/src/admin_cli/status.rs": 742,
-    "crates/rginx-app/src/main.rs": 842,
     "crates/rginx-config/src/compile/server.rs": 728,
     "crates/rginx-config/src/compile/upstream.rs": 510,
     "crates/rginx-config/src/model.rs": 673,
